Use Doxyfile to document libprimis-headers.
Modify the Doxyfile settings to include the correct project name and
description. Also, specify the files to document. The following files
and folders will be documented:
  - src/libprimis-headers/
  - README.md (this will be the main page)
  - doc/

The documentation will be saved in the docs folder, where GitHub pages
can read it.

LaTeX is disabled, and MathJax is enabled for math rendering.

Finally, a small SVG logo was included to decorate the Doxgyen
documentation site. The original icon file was too large and looked
pixelated when zooming in.