Create Doxygen Workflow.
The workflow checks out the current repository, runs Doxygen, and then
uploads the generated documentation folder to the gh-pages branch.

The GitHub token is necessary to authenticate pushing files into a
gh-pages branch.

The workflow uses the `edge` version of doxygen-action to always use
latest version of Doxygen.

Checking out the submodules is necessary so that Doxygen will be able to
document the libprimis-subheaders.